+++ tests/integration/condition_each_test.py
@@ -19,6 +19,8 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+import pytest
+
 from selene import have
 from selene.core.condition import Condition
 from selene.core.exceptions import TimeoutException
@@ -74,4 +76,53 @@ def test_on_collection(session_browser):
             "Reason: AssertionError: "
             "Not matched elements among all with indexes from 0 to 2:\n"
             "browser.all(('css selector', 'li')).cached[1]: condition not matched"
+        ) in str(error)
+
+
+def test_on_collection_with_expected_size(session_browser):
+    browser = session_browser.with_(timeout=0.1)
+    GivenPage(browser.driver).opened_with_body(
+        '''
+        <ul>Welcome to:
+           <li>Harry from Hogwarts</li>
+           <li>Ron from Hogwarts</li>
+           <li>Hermione from Hogwarts</li>
+        </ul>
+        '''
+    )
+
+    # WHEN
+    elements = browser.all('p')  # the size of elements collection is 0
+
+    # THEN this passes, because each among 0 - technically has 'from Hogwarts' :D
+    elements.should(have.text('from Hogwarts').each)
+    # AND all these too, because the STOP index is implicit so assume 0 length too
+    elements[:].should(have.text('from Hogwarts').each)
+    elements[0:].should(have.text('from Hogwarts').each)
+    elements[0:-1].should(have.text('from Hogwarts').each)
+
+    # BUT this DOES NOT:
+    try:
+        elements[:3].should(have.text('from Hogwarts').each)
+        pytest.fail("should have failed on size mismatch")
+    except TimeoutException as error:
+        assert (
+            "browser.all(('css selector', 'p'))[:3]. each has text from Hogwarts\n"
+            '\n'
+            'Reason: AssertionError: not enough elements to slice collection from START '
+            'to STOP at index=3, actual elements collection length is 0\n'
+        ) in str(error)
+
+    # AND while this pass
+    browser.all('li')[2:].should(have.text('from Hogwarts').each)
+    # BUT this DOES NOT too:
+    try:
+        browser.all('li')[3:].should(have.text('from Hogwarts').each)
+        pytest.fail("should have failed on size mismatch")
+    except TimeoutException as error:
+        assert (
+            "browser.all(('css selector', 'li'))[3:]. each has text from Hogwarts\n"
+            '\n'
+            'Reason: AssertionError: not enough elements to slice collection from START '
+            'on index=3, actual elements collection length is 3\n'
         ) in str(error)